UI: [#35] Fixing issues.
*Send UserInterface::InterfaceWithData in the queue instead of Event.
*Remove pointer madness for LCD.
*Timeout for Idle Event from Master thread.

+++ source/shoh/src/threads/manager/Menu.h
@@ -27,7 +27,7 @@ private:
   /* Variables and objects */
   ThreadCommon::QueueManager* _qm;
   p_state current;
-  int ext_temp;
+  short ext_temp;
   Counter<EventRawData> set_point;
   const char main_text[64];
   const char set_point_text[64];
